feat(mod): modulized the entire codebase
- moved codes to individual files
	- moved `BASEURL` to `r#static` module.
		- gived it a `public(crate)` scope.
	- moved `NekosLifeError` to `error` module.
	- moved `implementation` module to separated file.
		- added `#[path]` attribute on the blocking module.
		- moved blocking module code to `blocking.rs` file.
	- moved each test code under the `lib.rs` file to individual
	module files.
